fix: 64-bit thread-range split in the resident near-field build
The host classification loop split its pair range with

    int begin = (sys->nNearPairsFlat * t) / nThreads;

computing the product in int. Above INT_MAX/nThreads -- 67,108,863 pairs
at the 32-thread cap -- that overflows, giving a negative begin and end
and an out-of-range read of nearPairSrc. The result was SIGSEGV, with no
diagnostic.

It needed a case both large enough to overflow and small enough to stay
resident, and only one configuration lands in that gap: 6CO8 sdens=1 at
depth 9 has 92,779,203 near pairs (x32 = 2.97e9) and takes the resident
path, where the same mesh at the default depth 8 has 25.5M pairs and
streams instead. Every capsid run at default settings streams and never
enters this loop; every resident case so far was small. That is why it
survived until a depth study went looking.

Fixed with the (long long) cast used for the same split everywhere else
in the tree. An audit found two more instances of the identical pattern:
the preconditioner disjoint build in gpu_backend_cuda.cu, and the dense
direct baseline in gpu_direct_cuda.inc, which splits on nPnls and would
therefore overflow on any H1N1 sdens=1 run. Both fixed; no unguarded
splits remain.

Verified: 6CO8 sdens=1 depth 9 now completes -- 608,284,440 resident
panel pairs, 92 iterations, residual 9.899505e-05, energy
-123,468.878004, no fallback -- where it previously died at 2m44s.
1cbn/1ajj/1a63 are bit-identical.

+++ src/gpu_backend_cuda.cu
@@ -486,8 +486,9 @@ int gpuDirectApply(ssystem *sys, double alpha, double beta, const double *sgm, d
 
           workers.reserve((size_t)nThreads);
           for (t = 0; t < nThreads; t++) {
-            int begin = (dstCount * t) / nThreads;
-            int end = (dstCount * (t + 1)) / nThreads;
+            /* 64-bit split; see the note in gpu_nearfield_cuda.inc. */
+            int begin = (int)(((long long)dstCount * t) / nThreads);
+            int end = (int)(((long long)dstCount * (t + 1)) / nThreads);
             workers.emplace_back([=]() {
               int localDst, localS;
               for (localDst = begin; localDst < end; localDst++) {
